api: perform quota checks before certain actions
- Creating workflows
- Starting workflows
- Uploading files
- Opening interactive sessions

closes #282
closes #283
